+++ CHANGELOG.md
@@ -5,12 +5,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+--
+
+---
+
+### [v0.2.0](https://github.com/decodelabs/telegraph/commits/v0.2.0) - 13th June 2025
+
 - Setup Veneer frontage
 - Build Config interface and Dovetail implementation
 - Added initial Adapter interface structure
 - Added initial Source container
 - Implemented remote item models
 - Added SubscriptionRequest structure
+
+[Full list of changes](https://github.com/decodelabs/telegraph/compare/v0.1.1...v0.2.0)
 
 ---